feat(*): address review comments
